% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktion til at gemme.R
\encoding{UTF-8}
\name{gemDatasaetIDb}
\alias{gemDatasaetIDb}
\title{Gem datasæt i en database}
\usage{
gemDatasaetIDb(
  datasaet,
  schemaNavn,
  datasaetNavn = NA,
  databaseNavn = minDatabase,
  serverNavn = minServer,
  driverNavn = minDriver
)
}
\arguments{
\item{datasaet}{en data.frame, datasættet som skal gemmes}

\item{schemaNavn}{en character string, navnet på det schema datasættet skal gemmes i}

\item{datasaetNavn}{en character string, navnet som datasættet skal have i databasen. Navnet skal indeholde mindst en underscore "_".
Default er navnet på 'datasaet' og '_df'. Fx. hvis datasaet hedder 'aktivitet' vil datasættet hedde 'aktivitet_df' i databasen.}

\item{databaseNavn}{en character string, Navnet på databasen som datasættet skal gemmes i. Defaultværdien for variablen er værdien 'minDatabase' fra den globale environment.}

\item{serverNavn}{en character string, Navnet på serveren. Defaultværdien for variablen er værdien 'minServer' fra den globale environment.}

\item{driverNavn}{en character string, Navnet på driveren. Defaultværdien for variablen er værdien 'minDriver' fra den globale environment.}
}
\value{
ingen return
}
\description{
Funktionen bruges til at gemme et datasæt i en database
}
