% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indlaesningsfunktioner.R
\encoding{UTF-8}
\name{hentDatasaetFraDatabase}
\alias{hentDatasaetFraDatabase}
\title{Hent datasæt fra database}
\usage{
hentDatasaetFraDatabase(
  schemaNavn,
  datasaetNavn,
  datasaetFormatFil = NA,
  udtraeksDatoTidVaerdi = NA,
  hentForSidsteUdtraeksDatoTidVaerdi = FALSE,
  databaseNavn = minDatabase,
  serverNavn = minServer,
  driverNavn = minDriver
)
}
\arguments{
\item{schemaNavn}{en character string, Navnet på det schema tabellen ligger i i databasen}

\item{datasaetNavn}{en character string, Navnet på tabellen, som skal hentes}

\item{datasaetFormatFil}{en data.frame, formatfilen som fortæller, hvad og hvordan datasættet skal formateres (default værdi er NA - datasættet formateres ikke)}

\item{udtraeksDatoTidVaerdi}{en character string, angiver version for at udvælge en bestemt version af en tabel der opdateres løbende uden erstatning af tidligere versioner af datasættet}

\item{hentForSidsteUdtraeksDatoTidVaerdi}{TRUE/FALSE værdi, Hvis TRUE, så hentes de rækker af datasættet der svarer til den sidste udtrækdato. Kræver i øvrigt at 'udtraeksDatoTidVaerdi' er NA.}

\item{databaseNavn}{en character string, Navnet på databasen som datasættet skal hentes fra. Defaultværdien for variablen er værdien 'minDatabase' fra den globale environment.}

\item{serverNavn}{en character string, Navnet på serveren som datasættet skal hentes fra. Defaultværdien for variablen er værdien 'minServer' fra den globale environment.}

\item{driverNavn}{en character string, Navnet på driveren som skal bruges til at hente datasættet. Defaultværdien for variablen er værdien 'minDriver' fra den globale environment.}
}
\value{
Den indlæste datasæt
}
\description{
Funktion til at indlæse et datasæt fra database med følgende muligheder udover selve indlæsningen:\cr
    \itemize{
    \item Mulighed for at definere 'udtraeksDatoTidVaerdi' for at hente en bestemt version af et datasæt, der indeholder flere versioner af samme data.
    \item Mulighed for at formatere datasæt med det samme ved angivelse af datasaetFormatFil
    }
    Funktionen bruger værdier fra global environment til at finde databasen, serveren og driveren, der skal bruges til at hente data fra.
}
\seealso{
\code{\link[Formateringsfunktioner]{retOgFormaterDatasaet}} og \code{\link[Formateringsfunktioner]{opsaetFormatFil}}
}
