% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andre funktioner.R
\encoding{UTF-8}
\name{tjekOmSchemaEksistererIDb}
\alias{tjekOmSchemaEksistererIDb}
\title{Tjek om schema eksister i databasen}
\usage{
tjekOmSchemaEksistererIDb(
  schemaNavn,
  databaseNavn = minDatabase,
  serverNavn = minServer,
  driverNavn = minDriver
)
}
\arguments{
\item{schemaNavn}{en character string, navnet på det schema der skal tjekkes}

\item{databaseNavn}{en character string, Navnet på databasen for schema. Defaultværdien for variablen er værdien 'minDatabase' fra den globale environment.}

\item{serverNavn}{en character string, Navnet på serveren. Defaultværdien for variablen er værdien 'minServer' fra den globale environment.}

\item{driverNavn}{en character string, Navnet på driveren. Defaultværdien for variablen er værdien 'minDriver' fra den globale environment.}
}
\value{
Funktionen melder fejl, hvis schema ikke eksisterer i databasen ellers ingen return.
}
\description{
Funktionen bruges til at tjekke om et schema eksisterer i databasen
}
