% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andre funktioner.R
\encoding{UTF-8}
\name{tjekOmTabelEksistererIDb}
\alias{tjekOmTabelEksistererIDb}
\title{Tjek om tabellen eksister i databasen}
\usage{
tjekOmTabelEksistererIDb(
  tabelNavn,
  schemaNavn,
  databaseNavn = minDatabase,
  serverNavn = minServer,
  driverNavn = minDriver
)
}
\arguments{
\item{tabelNavn}{en character string, navnet på det tabel der skal tjekkes}

\item{schemaNavn}{en character string, navnet på det schema som tabellen skal findes i}

\item{databaseNavn}{en character string, Navnet på databasen. Defaultværdien for variablen er værdien 'minDatabase' fra den globale environment.}

\item{serverNavn}{en character string, Navnet på serveren. Defaultværdien for variablen er værdien 'minServer' fra den globale environment.}

\item{driverNavn}{en character string, Navnet på driveren. Defaultværdien for variablen er værdien 'minDriver' fra den globale environment.}
}
\value{
TRUE/FALSE, TRUE hvis tabellen findes i databasen
}
\description{
Funktionen bruges til at tjekke om en tabel eksisterer under en bestemt schema i databasen
}
